ensure consistent reads of tuf_artifact table

+++ schema/crdb/tuf-generation/up01.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS omicron.public.tuf_generation (
     singleton BOOL NOT NULL PRIMARY KEY,
-    generation INT NOT NULL,
+    generation INT8 NOT NULL,
     CHECK (singleton = true)
 );
 INSERT INTO omicron.public.tuf_generation (
